+++ packages/xrpl/HISTORY.md
@@ -5,6 +5,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 ## Unreleased
 ### Fixed
 * Client.disconnect() now stops the heartbeat health check as well
+* Updated dependencies which had a warning when running `npm audit`
 * Infinite error/reconnect in browser if an exception was raised during the initial websocket connection event.
 * Errors during reliable submission with no error message now properly show the preliminary result instead of a type error